pyproject.toml
[project]
name = "gemini-structured-output"
version = "0.1.0"
description = "D&D weapon generator using Gemini structured output with a Pydantic schema"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "google-genai>=1.0.0",
]
